requirements.txt
torch
torchdata
transformers
evaluate
rouge_score
loralib
peft
datasets
scikit-learn
langchain
langchain-core
langchain-community
python-dotenv
langchain-huggingface
streamlit
pydantic
evaluate